refactor(sdk-client): restructure logger implementation by introducing logger-types and separating SinchLogger functionality
