+++ packages/sdk-client/src/logger/index.ts
@@ -1,1 +1,7 @@
-export * from './logger';
+export type { LogMessage, Logger } from './logger-types';
+export {
+  CONSOLE_LOGGER,
+  NOOP_LOGGER,
+  resolveClientParameters,
+  resolveLogger,
+} from './logger';
