+++ packages/sdk-client/src/logger/logger-types.ts
@@ -1,0 +1,8 @@
+export type LogMessage = string | (() => string);
+
+export interface Logger {
+  debug(message: LogMessage, ...meta: any[]): void;
+  info(message: LogMessage, ...meta: any[]): void;
+  warn(message: LogMessage, ...meta: any[]): void;
+  error(message: LogMessage, ...meta: any[]): void;
+}
